+++ runner.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"io"
 	"io/ioutil"
-	"runtime/debug"
 	"sync"
 )
 
@@ -82,26 +81,7 @@ func (r taskRunner) run(in Input) Result {
 		logger: logger,
 	}
 
-	var (
-		ch         = make(chan struct{})
-		finished   bool
-		panicVal   interface{}
-		panicStack []byte
-	)
-	go func() {
-		defer close(ch)
-		defer a.runCleanups(&finished, &panicVal, &panicStack)
-		defer func() {
-			if finished {
-				return
-			}
-			panicVal = recover()
-			panicStack = debug.Stack()
-		}()
-		r.action(a)
-		finished = true
-	}()
-	<-ch
+	finished, panicVal, panicStack := a.run(r.action)
 
 	res := Result{}
 	switch {
@@ -109,7 +89,7 @@ func (r taskRunner) run(in Input) Result {
 		res.Status = StatusFailed
 	case a.Skipped():
 		res.Status = StatusSkipped
-	case finished && panicVal == nil:
+	case finished:
 		res.Status = StatusPassed
 	default:
 		res.Status = StatusFailed
@@ -117,52 +97,6 @@ func (r taskRunner) run(in Input) Result {
 		res.PanicStack = panicStack
 	}
 	return res
-}
-
-func (a *A) runCleanups(finished *bool, panicVal *interface{}, panicStack *[]byte) {
-	// we capture only the first panic
-	cleanupFinished := false
-	if *finished {
-		defer func() {
-			if cleanupFinished {
-				return
-			}
-			*panicVal = recover()
-			*panicStack = debug.Stack()
-			*finished = false
-		}()
-	} else {
-		defer func() {
-			_ = recover() // ignore next panics
-		}()
-	}
-
-	// Make sure that if a cleanup function panics,
-	// we still run the remaining cleanup functions.
-	defer func() {
-		a.mu.Lock()
-		recur := len(a.cleanups) > 0
-		a.mu.Unlock()
-		if recur {
-			a.runCleanups(finished, panicVal, panicStack)
-		}
-	}()
-
-	for {
-		var cleanup func()
-		a.mu.Lock()
-		if len(a.cleanups) > 0 {
-			last := len(a.cleanups) - 1
-			cleanup = a.cleanups[last]
-			a.cleanups = a.cleanups[:last]
-		}
-		a.mu.Unlock()
-		if cleanup == nil {
-			cleanupFinished = true
-			return
-		}
-		cleanup()
-	}
 }
 
 type syncWriter struct {
